+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 * st-select: Fix bug when select is used without being part of a form
 * st-table: Fix wrong css variable name
+* st-pagination: Remove duplicated event emitted when items per page are changed
 
 
 ## 21.0.0 (June 01, 2020)